refactor(codex): isolate app-server RPC errors
