+++ extensions/codex/src/app-server/client.ts
@@ -20,6 +20,7 @@ import {
   type RpcRequest,
   type RpcResponse,
 } from "./protocol.js";
+import { CodexAppServerRpcError } from "./rpc-error.js";
 import { createStdioTransport } from "./transport-stdio.js";
 import { createWebSocketTransport } from "./transport-websocket.js";
 import {
@@ -66,20 +67,7 @@ export function resolveCodexAppServerClientInstanceId(client: object): string {
   return getInstanceId?.call(client) ?? getCodexAppServerClientInstanceId(client);
 }
 
-/** RPC error wrapper that preserves app-server error code and data. */
-export class CodexAppServerRpcError extends Error {
-  readonly code?: number;
-  readonly data?: JsonValue;
-  readonly method: string;
-
-  constructor(error: { code?: number; message: string; data?: JsonValue }, method: string) {
-    super(formatCodexAppServerRpcErrorMessage(error, method));
-    this.name = "CodexAppServerRpcError";
-    this.code = error.code;
-    this.data = error.data;
-    this.method = method;
-  }
-}
+export { CodexAppServerRpcError } from "./rpc-error.js";
 
 class CodexAppServerLocalRequestCancellationError extends Error {
   readonly code = "CODEX_APP_SERVER_LOCAL_REQUEST_CANCELLED";
@@ -165,32 +153,6 @@ export function isCodexAppServerIndeterminateTransportError(error: unknown): err
     "mayHaveWritten" in error &&
     error.mayHaveWritten === true
   );
-}
-
-function formatCodexAppServerRpcErrorMessage(
-  error: { message: string; data?: JsonValue },
-  method: string,
-): string {
-  const message = error.message || `${method} failed`;
-  const detail = readCodexAppServerRpcReloginDetail(error.data);
-  return detail && !message.includes(detail) ? `${message}: ${detail}` : message;
-}
-
-function readCodexAppServerRpcReloginDetail(data: JsonValue | undefined): string | undefined {
-  const record = isJsonObject(data) ? data : undefined;
-  const nested = isJsonObject(record?.error) ? record.error : record;
-  if (!nested) {
-    return undefined;
-  }
-  const isRelogin =
-    nested.action === "relogin" ||
-    (nested.reason === "cloudRequirements" && nested.errorCode === "Auth");
-  const detail = typeof nested.detail === "string" ? nested.detail.trim() : "";
-  return isRelogin && detail ? detail : undefined;
-}
-
-function isJsonObject(value: unknown): value is { [key: string]: JsonValue } {
-  return Boolean(value && typeof value === "object" && !Array.isArray(value));
 }
 
 /** Returns true for errors that mean the app-server transport is closed. */
